requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-genai>=0.3.0
openai>=1.12.0
pillow>=10.0.0
requests>=2.31.0
